Refactor key state handling to use uint8_t for key codes

+++ include/plugin/Events.hpp
@@ -1,25 +1,24 @@
 #pragma once
 
+#include <cstdint>
 #include <variant>
 
 #include <Windows.h>
 
-struct OnPluginStart {
-};
+struct OnPluginStart {};
 
-struct OnPluginEnd {
-};
+struct OnPluginEnd {};
 
 struct OnKeyDown {
-    const int vKey;
+    const uint8_t vKey;
 };
 
 struct OnKeyHold {
-    const int vKey;
+    const uint8_t vKey;
 };
 
 struct OnKeyUp {
-    const int vKey;
+    const uint8_t vKey;
 };
 
 struct OnCursorVisibilityChange {
